+++ FileSystems/Users.js
@@ -4,16 +4,12 @@ let dir = '';
 
 export const mkUser = (id) => {
     dir = 'Store/Users/' + id + '/Invoices/';
-    if (!fs.existsSync(dir)) {
-        fs.mkdirSync(dir, { recursive: true });
-    }
+    if (!fs.existsSync(dir)) fs.mkdirSync(dir, { recursive: true });
 }
 
 /*
 export const rmUser = (id) => {
     dir = 'Store/Users/' + id;
-    if (fs.existsSync(dir)) {
-        fs.rmSync(dir, { recursive: true });
-    }
+    if (fs.existsSync(dir)) fs.rmSync(dir, { recursive: true });
 }
 */
